requirements.txt
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cache==0.12
Flask-CKEditor==0.4.6
Flask-DotEnv==0.1.2
flask-heroku==0.1.9
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-mongoengine==1.0.0
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.10
Flask-Script==0.6.7
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.4.1
Flask-Uploads==0.1.3
Flask-WTF==1.1.1
Werkzeug==2.2.2